siamtemp_hvac_chatbot/agents/parallel/processor.py
import os
import re
import json
import asyncio
import time
import logging
import hashlib
from typing import Dict, List, Any, Optional, Tuple, Union
from datetime import datetime, date, timedelta
from decimal import Decimal
from collections import deque, defaultdict
import aiohttp
import psycopg2
from textwrap import dedent
from psycopg2.extras import RealDictCursor
from collections import Counter, defaultdict
logger = logging.getLogger(__name__)

class ParallelProcessingEngine:
    """
    ระบบประมวลผลแบบขนาน - ยังไม่ได้ implement เต็มรูปแบบ
    แต่เตรียมโครงสร้างไว้สำหรับอนาคต
    """
    def __init__(self):
        self.performance_stats = defaultdict(list)
    
    async def parallel_analyze(self, question: str, context: Dict) -> Dict[str, Any]:
        """วิเคราะห์แบบขนาน (สำหรับอนาคต)"""
        # ตอนนี้ยังเป็น sequential
        return {
            'intent': 'sales',
            'entities': {},
            'complexity': 'simple'
        }