siamtemp_hvac_chatbot/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
asyncpg
# Database
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP Client
aiohttp==3.9.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
prometheus-client==0.19.0

# Logging
loguru==0.7.2